src/backend/synapse/requirements.txt
anyio==3.6.2
click==8.1.6
fastapi==0.95.1
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.23.1
pynest-api==0.1.0
    